+++ Assets/Neoxider/Scripts/Tools/View/MeshEmission.cs
@@ -121,7 +121,14 @@ namespace Neo.Tools.View
         {
             if (_material != null)
             {
-                DestroyImmediate(_material);
+                if (Application.isPlaying)
+                {
+                    Destroy(_material);
+                }
+                else
+                {
+                    DestroyImmediate(_material);
+                }
                 _material = null;
             }
         }